+++ src/AndreyNaumov/Test.java
@@ -1,0 +1,16 @@
+package AndreyNaumov;
+
+public class Test {
+    public static void main(String[] args) {
+        System.out.println("FINRA task");
+        util.finra();
+        System.out.println();
+        System.out.println("------------------------------------------------------------------------");
+        util.finra1();
+        System.out.println();
+        System.out.println();
+
+        Math.round()
+
+    }
+}
