+++ src/AndreyNaumov/numbers_odd_or_even.java
@@ -8,4 +8,7 @@ identify(5); ->"Odd"
 identify(6); ->"Even"
  */
 public class numbers_odd_or_even {
+
+
+
 }
